Refactor server to avoid globals
I didn't like the "magic" of calling e.g. `InitDb()` in `main` then
using `GetDB()` in handlers; it set up an implicit dependency. I
switched it to use a custom Echo context initialized and passed
explicitly to handlers. The `CheckinStore` is testable as well; we can
pass mocks to `NewCheckinStore()`.